Added tagging support

+++ yukon-client/tdc/src/com/cannontech/tdc/utils/TDCDefines.java
@@ -29,7 +29,7 @@ public final class TDCDefines
 	
 	// JTable model constants
 	public static final int ROW_BREAK_COUNT = 8;
-	public static final long ROW_BREAK_ID = 
+	public static final int ROW_BREAK_ID = 
             com.cannontech.database.data.point.PointTypes.SYS_PID_SYSTEM;
 
 	public static int MAX_ROWS = 500;
